--- component/PageContainer/index.ts
@@ -1,1 +1,0 @@
-export { PageContainer } from "component/PageContainer/PageContainer";
