--- component/PageContent/index.ts
@@ -1,1 +1,0 @@
-export { PageContent } from "component/PageContent/PageContent";
