+++ components/layout/LayoutErrorBoundary.tsx
@@ -1,0 +1,16 @@
+import type { FC, ReactNode } from "react";
+import type { FallbackProps } from "react-error-boundary";
+import { ErrorBoundary } from "react-error-boundary";
+
+export const LayoutErrorBoundary: FC<{ children: ReactNode }> = ({
+  children
+}) => (
+  <ErrorBoundary FallbackComponent={ErrorFallback}>{children}</ErrorBoundary>
+);
+
+const ErrorFallback = ({ error }: FallbackProps) => (
+  <div>
+    <p>Something went wrong:</p>
+    <pre>{error.message}</pre>
+  </div>
+);
