--- layout/LayoutErrorBoundary.tsx
@@ -1,20 +1,0 @@
-import type { FC, ReactNode } from "react";
-import type { FallbackProps } from "react-error-boundary";
-import { ErrorBoundary } from "react-error-boundary";
-
-export const LayoutErrorBoundary: FC<{ children: ReactNode }> = ({
-  children,
-}) => {
-  return (
-    <ErrorBoundary FallbackComponent={ErrorFallback}>{children}</ErrorBoundary>
-  );
-};
-
-const ErrorFallback = ({ error }: FallbackProps) => {
-  return (
-    <div>
-      <p>Something went wrong:</p>
-      <pre>{error.message}</pre>
-    </div>
-  );
-};
