--- lib/const/index.ts
@@ -1,1 +1,0 @@
-export { getPath } from "lib/const/path";
