--- lib/mantine/index.ts
@@ -1,2 +1,0 @@
-export { AppMantineProvider } from "lib/mantine/AppMantineProvider";
-export { GlobalStyleProvider } from "lib/mantine/GlobalStyleProvider";
