+++ lib/provider/GlobalStyleProvider.tsx
@@ -1,0 +1,16 @@
+import { Global } from "@mantine/core";
+import type { FC, ReactNode } from "react";
+
+export const GlobalStyleProvider: FC<{ children: ReactNode }> = ({
+  children
+}) => (
+  <>
+    <Global
+      styles={{
+        "*": { wordBreak: "break-all" },
+        'button, *[type="button"]': { transform: "none !important" }
+      }}
+    />
+    {children}
+  </>
+);
