+++ pages/home.tsx
@@ -1,9 +1,9 @@
 import { useAuth0 } from "@auth0/auth0-react";
 import { Box, Button, Stack, Text } from "@mantine/core";
 import type { CustomNextPage } from "next";
-import { DashboardLayout } from "layout";
-import { PageContainer } from "component/PageContainer";
-import { PageContent } from "component/PageContent";
+import { Dashboard } from "components/layout/dashboard/Dashboard";
+import { PageContainer } from "components/PageContainer";
+import { ContentCard } from "components/ContentCard";
 
 const Home: CustomNextPage = () => {
   const { isAuthenticated, loginWithRedirect, logout, user } = useAuth0();
@@ -11,25 +11,24 @@ const Home: CustomNextPage = () => {
   return (
     <PageContainer title="ホーム" fluid>
       <Stack spacing="xl">
-        <PageContent title="ログイン">
-          {isAuthenticated && (
+        <ContentCard title="ログイン">
+          {isAuthenticated ? (
             <Box>
               <Text>{user?.name}でログイン中</Text>
               <Button onClick={() => logout()}>ログアウト</Button>
             </Box>
-          )}
-          {!isAuthenticated && (
+          ) : (
             <Box>
               <Text>ログインしていません。ログインしますか？</Text>
               <Button onClick={() => loginWithRedirect()}>ログイン</Button>
             </Box>
           )}
-        </PageContent>
+        </ContentCard>
       </Stack>
     </PageContainer>
   );
 };
 
-Home.getLayout = DashboardLayout;
+Home.getLayout = Dashboard;
 
 export default Home;
